实施 docs/design/abstract-model.md 三层 fold 结构:
- engine 拆 GameState → MatchState + RoundState (type-state) + AtomicOp
- stage 6/7 GameEngine wrapper 替换 UI/net/dev 层全部 GameState 引用
- net/p2p Phase A/B/D: 协议错误转 GameOver / 30s 重连 / Standard/ZeroTrust
- net/room 拆 2060 行单文件 → 6 子模块 (mod/lobby/game/projection/zerotrust/tests)
- ZeroTrust 不足 4 真人自动降级 Standard + AI 补足
- LobbyAnnouncement 同步真实玩家数 + lifecycle (修 host.rs:372 TODO)
- 测试覆盖率从 ~76% 推到 84.5%, 在线游戏 UI 三屏从 0% 到 88-98%

+++ src/net/protocol.rs
@@ -14,13 +14,13 @@
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
-use crate::domain::meld::{Meld, Seat};
-use crate::domain::tile::{Tile, TileIndex};
+use crate::engine::domain::meld::{Meld, Seat};
+use crate::engine::domain::tile::{Tile, TileIndex};
 use crate::engine::event::GameEvent;
 use crate::engine::phase::Phase;
+use crate::engine::round_state::RoundWind;
 use crate::engine::rules::GameRules;
 use crate::engine::score::Ranking;
-use crate::engine::state::RoundWind;
 use crate::ui::screens::game::TileSpec;
 
 // ============================================================================
@@ -178,7 +178,7 @@ pub struct GameStateView {
     pub round_wind: RoundWind,
     pub kyoku: u8,
     pub honba: u8,
-    pub riichi_sticks: u8,
+    pub riichi_sticks: u32,
     pub dealer: Seat,
     pub turn: Seat,
     pub phase: Phase,
